Remove the unused file
